Remove ActiveSupport requirement
I require ActiveSupport::Inflector to make use of the titleize method
but including this library monkey patches String and causes some
benchmarks to slow down. Replacing their titleize with a very simple
verion of my own. If you're not happy with the way the benchmark is
named, use the #name method after extending bench_press to specify the
name of the benchmark.